separated spec files for count-up
