+++ cypress/integration/v2/blocks/count-up/mobile-advanced.spec.js
@@ -1,0 +1,11 @@
+/**
+ * External dependencies
+ */
+import { registerTests } from '~stackable-e2e/helpers'
+
+/**
+ * Internal dependencies
+ */
+import { mobileAdvanced } from '.'
+
+describe( 'Count Up Block ( Mobile Advanced )', registerTests( [ mobileAdvanced ] ) )
