+++ cypress/integration/v2/blocks/count-up/tablet-style.spec.js
@@ -1,0 +1,11 @@
+/**
+ * External dependencies
+ */
+import { registerTests } from '~stackable-e2e/helpers'
+
+/**
+ * Internal dependencies
+ */
+import { tabletStyle } from '.'
+
+describe( 'Count Up Block ( Tablet Style )', registerTests( [ tabletStyle ] ) )
